+++ src/components/utils/arrIsEmpty.js
@@ -1,0 +1,5 @@
+import { isArray, isEmpty } from "lodash";
+
+const arrIsEmpty = (obj) => isEmpty(obj) || !isArray(obj);
+
+export { arrIsEmpty };