+++ src/components/utils/objIsEmpty.js
@@ -1,0 +1,5 @@
+import { isArray, isEmpty, isString } from "lodash";
+
+const objIsEmpty = (obj) => isEmpty(obj) || isArray(obj) || isString(obj);
+
+export { objIsEmpty };