+++ src/components/utils/objIsObject.js
@@ -1,0 +1,5 @@
+import { isObject, isArray } from "lodash";
+
+const objIsObject = (obj) => isObject(obj) && !isArray(obj) && typeof obj === 'object';
+
+export { objIsObject };